+++ public_transport_watcher/db/models/enums/day_category.py
@@ -1,0 +1,9 @@
+import enum
+
+
+class DayCategoryEnum(enum.Enum):
+    JOHV = "JOHV"  # business day outside school holidays
+    SAHV = "SAHV"  # saturday outside school holidays
+    JOVS = "JOVS"  # business day in school holidays
+    SAVS = "SAVS"  # saturday in school holidays
+    DIJFP = "DIJFP"  # sunday, bank holidays and other special days