+++ public_transport_watcher/db/models/geography.py
@@ -1,5 +1,6 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy import Column, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
+
 from public_transport_watcher.db.models.base import Base
 
 geography_schema = "geography"