+++ public_transport_watcher/extractor/configuration/mapping.py
@@ -1,0 +1,24 @@
+COLUMN_MAPPING = {
+    "navigo": {
+        "profils": [
+            "code_stif_trns",
+            "code_stif_res",
+            "code_stif_arret",
+            "libelle_arret",
+            "ida",
+            "cat_jour",
+            "trnc_horr_60",
+            "pourc_validations",
+        ],
+        "validations": [
+            "jour",
+            "code_stif_trns",
+            "code_stif_res",
+            "code_stif_arret",
+            "libelle_arret",
+            "ida",
+            "categorie_titre",
+            "nb_vald",
+        ],
+    }
+}