+++ public_transport_watcher/extractor/extract/__init__.py
@@ -1,0 +1,3 @@
+from .navigo import extract_navigo_validations
+
+__all__ = ["extract_navigo_validations"]