+++ public_transport_watcher/extractor/insert/__init__.py
@@ -1,0 +1,5 @@
+from .navigo import insert_navigo_data
+
+__all__ = [
+    "insert_navigo_data",
+]