+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-03-29
+MR #2
+
+### Added
+- Navigo validations data ingestion
+- Utils function to get the credentials, datalake files and engine
+- Logging configuration
+
+### Changed
+- Time bin table to include the type of day for transport data
+- Station base table to remove the "wording" column
+
 ## [0.1.0] - 2025-03-23
 MR #1